#281 Cleanup Sandhold
## Changes
- introduced construction-helper which handles the switching of prefabs
