+++ apps/client/src/app/probable-waffle/game/prefabs/buildings/tivara/Sandhold.ts
@@ -8,11 +8,7 @@ import SandholdFoundation1 from "./Sandhold/SandholdFoundation1";
 import SandholdFoundation2 from "./Sandhold/SandholdFoundation2";
 /* START-USER-IMPORTS */
 import { ObjectNames } from "../../../data/object-names";
-import { ActorDataChangedEvent } from "../../../data/actor-data";
-import { getActorComponent } from "../../../data/actor-component";
-import { ConstructionSiteComponent } from "../../../entity/building/construction/construction-site-component";
-import { Subscription } from "rxjs";
-import { environment } from "../../../../../../environments/environment";
+import { ConstructionHelper } from "../../../entity/building/construction/construction-helper";
 /* END-USER-IMPORTS */
 
 export default class Sandhold extends Phaser.GameObjects.Container {
@@ -57,15 +53,8 @@ export default class Sandhold extends Phaser.GameObjects.Container {
   /* START-USER-CODE */
   name = ObjectNames.Sandhold;
 
-  private constructionSiteHandlerSetup = false;
-  private constructionSubscription?: Subscription;
   private setup() {
-    this.on(ActorDataChangedEvent, this.actorDataChanged, this);
-    this.setupTestClickBuildEvent();
-  }
-
-  private actorDataChanged() {
-    this.constructionSiteHandler();
+    new ConstructionHelper(this, this.handlePrefabVisibility.bind(this));
   }
 
   private handlePrefabVisibility(progress: number | null) {
@@ -74,43 +63,6 @@ export default class Sandhold extends Phaser.GameObjects.Container {
     this.sandholdFoundation1.visible = progress !== null && progress < 50;
     this.sandholdFoundation2.visible = progress !== null && progress >= 50 && progress < 100;
   }
-
-  private constructionSiteHandler() {
-    if (this.constructionSiteHandlerSetup) return;
-    const constructionSiteComponent = getActorComponent(this, ConstructionSiteComponent);
-    if (!constructionSiteComponent) {
-      this.constructionSubscription?.unsubscribe();
-      this.handlePrefabVisibility(null);
-      return;
-    }
-    this.constructionSiteHandlerSetup = true;
-    if (constructionSiteComponent.isFinished()) {
-      this.handlePrefabVisibility(100);
-    } else {
-      this.constructionSubscription = constructionSiteComponent.constructionProgressPercentageChanged.subscribe(
-        (progress) => this.handlePrefabVisibility(progress)
-      );
-    }
-  }
-
-  private setupTestClickBuildEvent() {
-    if (environment.production) return;
-    console.log("Just test click build event - remove this");
-    this.sandholdFoundation1.on("pointerdown", () => {
-      const constructionSiteComponent = getActorComponent(this, ConstructionSiteComponent);
-      if (!constructionSiteComponent) return;
-      constructionSiteComponent.startConstruction();
-      constructionSiteComponent.assignBuilder(this); // todo
-      constructionSiteComponent.update(0, 200); // todo
-    });
-  }
-
-  destroy(fromScene?: boolean) {
-    this.off(ActorDataChangedEvent, this.actorDataChanged, this);
-    this.constructionSubscription?.unsubscribe();
-    super.destroy(fromScene);
-  }
-
   /* END-USER-CODE */
 }
